refactor: upgrade order and product identifiers to long and implement primary constructor syntax in controllers

+++ eCommerce.Storefront.Controllers/Controllers/PaymentController.cs
@@ -62,7 +62,7 @@ namespace eCommerce.Storefront.Controllers.Controllers
             }
         }
 
-        public async Task<IActionResult> CreatePaymentFor(int orderId)
+        public async Task<IActionResult> CreatePaymentFor(long orderId)
         {
             var request = new GetOrderRequest
             {
